refactor: 기존 결제 로직 리펙터링

+++ apps/commerce-api/src/main/kotlin/com/loopers/application/point/PointInfo.kt
@@ -1,10 +1,11 @@
 package com.loopers.application.point
 
 import com.loopers.domain.point.Point
+import java.math.BigDecimal
 
 data class PointInfo(
     val userId: Long,
-    val amount: Int,
+    val amount: BigDecimal,
 ) {
     companion object {
         fun from(point: Point?): PointInfo? {
@@ -12,18 +13,18 @@ data class PointInfo(
                 ?.let { PointInfo(point.userId, point.amount.value) }
         }
 
-        fun of(userId: Long, amount: Int): PointInfo {
+        fun of(userId: Long, amount: BigDecimal): PointInfo {
             return PointInfo(userId, amount)
         }
     }
 
     data class Charge(
         val userName: String,
-        val amount: Int,
+        val amount: BigDecimal,
     ) {
 
         companion object {
-            fun of(userName: String, amount: Int): Charge {
+            fun of(userName: String, amount: BigDecimal): Charge {
                 return Charge(userName, amount)
             }
         }
