+++ apps/commerce-api/src/main/kotlin/com/loopers/interfaces/api/point/PointV1Dto.kt
@@ -2,11 +2,12 @@ package com.loopers.interfaces.api.point
 
 import com.loopers.application.point.PointInfo
 import jakarta.validation.constraints.Min
+import java.math.BigDecimal
 
 class PointV1Dto {
     data class PointResponse(
         val userId: Long,
-        val amount: Int,
+        val amount: BigDecimal,
     ) {
         companion object {
             fun from(point: PointInfo?): PointResponse? {
@@ -18,7 +19,7 @@ class PointV1Dto {
 
     data class ChargeRequest(
         @field:Min(value = 0, message = "0 이하")
-        val amount: Int,
+        val amount: BigDecimal,
     ) {
         fun toCharge(userName: String): PointInfo.Charge {
             return PointInfo.Charge.of(userName, amount)
